requirements.txt
appdirs==1.4.4
asgiref==3.4.1
backcall==0.2.0
black==21.7b0
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
decorator==5.0.9
Django==3.2.5
django-filter==2.4.0
djangorestframework==3.12.4
httplib2==0.19.1
huey==2.3.2
idna==3.2
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
matplotlib-inline==0.1.2
mypy-extensions==0.4.3
packaging==21.0
parso==0.8.2
pathspec==0.9.0
pickleshare==0.7.5
prompt-toolkit==3.0.19
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.9.0
pyparsing==2.4.7
pytz==2021.1
redis==3.5.3
regex==2021.7.6
requests==2.26.0
rq==1.9.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
tomli==1.0.4
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.26.6
wcwidth==0.2.5
